pyproject.toml
[project]
name = "ai-frontend-gui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "pulp>=3.3.0",
    "pygame-ce>=2.5.7",
    "pywebview>=6.2.1",
]
